Switch to using `mb_encode_mimeheader` to encode our headers
This method properly handles creating encoded words, ensuring that we don’t split multi-byte characters between encoded words (RFC 2047 section 3), and ensuring that headers (including the header name) do not exceed 75 characters per line.

Involves a significant (and breaking) change to modify the public `encodeHeaderValue` method to be `encodeHeader`, so that the encode method is able to take into account the header name length when keeping the header values to 75 characters per line.