Create integration tests for testing using the outputted mail string as input to the mail factory
This includes a breaking test which highlights an issue with having no trailing new lines in the content part of a content only mail string. Due to a bug in mailparse (https://bugs.php.net/bug.php?id=75923), the last line of the mail string is discarded if it does not end with a trailing new line.