Get decoded/parsed headers from `zbateson/mail-mime-parser`
Currently, we're decoding the headers ourself, and parsing the email addresses using the mailparse extension (which we ideally want to completely move away from).

All of this is done by the mail parser library if we get the headers as header objects rather than getting the raw header values.

This does mean removing the 2 deprecated methods, so this will need to be a new major version.