Add performance tests
Adds a test which demonstrates a memory leak issue with the current implementation (due to this bug https://bugs.php.net/bug.php?id=74215 in the mailparse extension).

Also adds a test which sets a baseline target for parsing speed of the factory (to ensure that if we move away from mailparse, we still aim to maintain a certain level of performance).

These tests are very slow, and also quite subjective to environment, so I have set them to be disabled by default, only enabled when the `RUN_PERFORMANCE_TESTS` global variable is set to true for the unit tests.