Solve issues with content type not being set on a part
The code was previously sort of handling the case where we don't get a content type when parsing a mime mail, but this had some issues. There were parts which were defining a default content type, but the usage in the factory meant that it would never use the argument defaults (as it would always pass in a value for the argument, and this would be `false` if we were unable to determine a content type).

Change this to default to `null` in the factory, and ensure that explicitly passing in `null` for the `type` argument will use the default content type.

Added tests to ensure this behaviour is defined.