Store and build all headers using Symfony/Mime
Headers are stored internally as objects, which allows any exceptions for invalid values to be thrown when added rather than on output.
Main encode method is now `final`, and calls a build method to allow child classes to add additional headers.
This is so the main method can handle creating the Headers object and returning the final string,
but the child classes can add their headers before or after the parent as they need.
Attachment test is updated to use a regex to allow any whitespace (including new-lines) in the Content-Type header.